Separate CDC specific stuff from generic Kafka Connect source
